Time: 4 ms (42.97%), Space: 69.5 MB (80.41%) - LeetHub

+++ 0643-maximum-average-subarray-i/0643-maximum-average-subarray-i.java
@@ -7,6 +7,6 @@ class Solution {
             a=a-nums[i-k]+nums[i];
             b=Math.max(b, a);
         }
-        return b/k;
+        return (double) b/k;
     }
 }
